[TS-3735]: Coverity CID 1309075 - Incorrect expression.

+++ iocore/net/SSLNetVConnection.cc
@@ -466,7 +466,7 @@ SSLNetVConnection::net_read_io(NetHandler *nh, EThread *lthread)
       readSignalError(nh, err);
     } else if (ret == SSL_HANDSHAKE_WANT_READ || ret == SSL_HANDSHAKE_WANT_ACCEPT || ret == EVENT_CONT) {
       if (SSLConfigParams::ssl_handshake_timeout_in > 0) {
-        double handshake_time = ((Thread::get_hrtime() - sslHandshakeBeginTime) / 1000000000);
+        double handshake_time = ((double)(Thread::get_hrtime() - sslHandshakeBeginTime) / 1000000000);
         Debug("ssl", "ssl handshake for vc %p, took %.3f seconds, configured handshake_timer: %d", this, handshake_time,
               SSLConfigParams::ssl_handshake_timeout_in);
         if (handshake_time > SSLConfigParams::ssl_handshake_timeout_in) {
